+++ README.md
@@ -21,10 +21,11 @@ Remove each item from the list as you complete it.
 ## Design Setup
 
 1. Add any typography to `app/styles/fonts/typography.ts`, and configure a default text color/style in `app/styles/project.tsx`
-1. Add any text styles to `app/styles/text.ts`
+1. Add any text styles to `app/styles/text.ts` — get capsize font metrics from [seek-oss.github.io/capsize](https://seek-oss.github.io/capsize/)
 1. Configure the library config according to this project's needs.
 1. Update the `app/manifest.webmanifest` to include colors and metadata
 1. If they exist, export the fullmark, logomark, and wordmark from figma and add them to `app/images`
+1. Replace `app/favicon.ico`, `app/opengraph-image.jpg`, and `app/twitter-image.jpg` with project assets
 
 ## Sanity Setup or Removal
 
@@ -33,7 +34,7 @@ If this project is using a CMS, set up a new project in Sanity. If we're not sur
 ### Removing Sanity
 
 1. in `package.json`, delete any scripts that reference sanity.
-1. delete `app/(sanity)`, `app/blog`, `app/[[...slug]]`, and `sanity` folders
+1. delete `app/(sanity)`, `app/(blog-templates)`, `app/[[...slug]]`, and `sanity` folders
 1. delete `.vscode/tasks.json`, `sanity.config.ts`, `sanity.types.ts`, and `sanity.cli.ts`, and any other sanity-related files in the root of the project
 1. remove `SanityLive`, `SanityFetch`, and `defineQuery`, as well as any dependent code, from `app/layout.tsx`, `app/components/Header.tsx`, `app/components/Footer.tsx`, and `app/sitemap.ts`
 1. add `library/sanity` and `library/UniversalImage.tsx` to the `exclude` array in `tsconfig.json` (you'll be using StaticImage instead)
@@ -49,6 +50,15 @@ If this project is using a CMS, set up a new project in Sanity. If we're not sur
    - `NEXT_PUBLIC_SANITY_PROJECT_ID` is your project ID, for example `m85xxx23`
    - `NEXT_PUBLIC_SANITY_DATASET` is your dataset name, for a new project this is `production` (if we're working on updates while the site is live we'll use `development`)
    - `SANITY_AUTH_TOKEN` is a read only access token. Generate one in Sanity under the `API` tab
+
+## Blog Template Adoption
+
+If this project uses a blog template, adopt it from the starter by completing these steps. Full details in `sanity/schemas/blog/blog-1/README.md`.
+
+1. Set the hub slug in Sanity Studio (Studio → Blog 1 → Hub Settings) — this drives the public URL, sitemap, and canonical URLs
+1. Replace `blog1*` placeholder color tokens in `app/styles/colors.css.ts` with project brand colors
+1. Update grid values in `app/(blog-templates)/(blog-template-1)/[blogSlug]/layout.tsx` to match the design
+1. Delete unused template folders under `app/(blog-templates)/` and `sanity/schemas/blog/`
 
 <!-- Repository Cover -->
 <!-- you can get a nice image by modifying the figma cover, or using the opengraph image -->
@@ -71,3 +81,7 @@ Built with Next. Install dependencies with `pnpm i` and run with `pnpm dev`
    Wireit by default will run tasks in parallel where possible. It suppresses logs when doing so, unless there's an error. If you want to see verbose logging output, you can set an environment variable to configure the behavior. You can also disable parallel runs entirely in a similar fashion. See their readme for more details: https://github.com/google/wireit#environment-variables
 
    I've configured my default logger to be `WIREIT_LOGGER=metrics`
+
+---
+
+See [CONVENTIONS.md](CONVENTIONS.md) for code and project structure conventions.
