+++ app/(blog-templates)/(blog-template-1)/[blogSlug]/(home)/BlogHomeClient.tsx
@@ -11,8 +11,8 @@ import {
 } from "app/(blog-templates)/(blog-template-1)/[blogSlug]/components/SearchBar"
 import SmallCard from "app/(blog-templates)/(blog-template-1)/[blogSlug]/components/SmallCard"
 import { useHeaderTheme } from "app/(blog-templates)/(blog-template-1)/[blogSlug]/components/HeaderTheme"
-import { colors } from "app/styles/colors.css"
-import textStyles from "app/styles/text"
+import { colors } from "styles/colors.css"
+import textStyles from "styles/text"
 import { css, f, styled } from "library/styled/alpha"
 import { useMedia } from "library/useMedia"
 import { useSearchResults } from "library/useSearchResults"
