+++ app/(blog-templates)/(blog-template-1)/[blogSlug]/(home)/page.tsx
@@ -1,4 +1,4 @@
-import { colors } from "app/styles/colors.css"
+import { colors } from "styles/colors.css"
 import { resolveOpenGraphImage } from "library/sanity/opengraph"
 import { css, f, styled } from "library/styled/alpha"
 import { siteURL } from "library/siteURL"
