+++ app/(blog-templates)/(blog-template-1)/[blogSlug]/components/ButtonLink/index.tsx
@@ -1,9 +1,9 @@
-import { colors } from "app/styles/colors.css"
-import textStyles from "app/styles/text"
+import { colors } from "styles/colors.css"
+import textStyles from "styles/text"
 import type { UniversalLinkProps } from "library/link"
 import UniversalLink from "library/link"
 import { css, f, styled } from "library/styled/alpha"
-import ArrowIcon from "images/icons/Arrow.inline.svg"
+import ArrowIcon from "./Arrow.inline.svg"
 
 export default function ButtonLink({
 	children,
