+++ app/(blog-templates)/(blog-template-1)/[blogSlug]/components/Categories.tsx
@@ -1,7 +1,7 @@
 "use client"
 
-import { colors } from "app/styles/colors.css"
-import textStyles from "app/styles/text"
+import { colors } from "styles/colors.css"
+import textStyles from "styles/text"
 import { css, f, styled } from "library/styled/alpha"
 import { useBlogCategory } from "./SearchBar"
 import type { Card } from "../types"
