+++ app/(blog-templates)/(blog-template-1)/[blogSlug]/components/LargeCard.tsx
@@ -3,8 +3,8 @@
 import { useRouter } from "next/navigation"
 import { useState } from "react"
 import ButtonLink from "./ButtonLink"
-import { colors } from "app/styles/colors.css"
-import textStyles, { clampText } from "app/styles/text"
+import { colors } from "styles/colors.css"
+import textStyles, { clampText } from "styles/text"
 import UniversalImage from "library/UniversalImage"
 import { css, f, styled } from "library/styled/alpha"
 import type { FeaturedCard } from "../types"
