+++ app/action-test/page.tsx
@@ -5,7 +5,7 @@ import { Field } from "@base-ui/react/field"
 import { Form } from "@base-ui/react/form"
 import { Button } from "@base-ui/react/button"
 import { css, f, styled } from "library/styled/alpha"
-import { colors } from "app/styles/colors.css"
+import { colors } from "styles/colors.css"
 import { submitForm, type FormState } from "./actions"
 
 export default function ActionTestForm() {
