+++ app/components/Form/Fields/CheckboxGroupField.tsx
@@ -8,7 +8,7 @@ import type { FieldRoot } from "@base-ui/react/field"
 import type { Form } from "@base-ui/react/form"
 import { Fieldset } from "@base-ui/react/fieldset"
 import { css, f, styled } from "library/styled/alpha"
-import { colors } from "app/styles/colors.css"
+import { colors } from "styles/colors.css"
 
 interface Option {
 	value: string | null
