+++ app/components/Form/Fields/SliderField.tsx
@@ -7,7 +7,7 @@ import type { FieldRoot } from "@base-ui/react/field"
 import type { Form } from "@base-ui/react/form"
 import { Fieldset } from "@base-ui/react/fieldset"
 import { css, f, styled } from "library/styled/alpha"
-import { colors } from "app/styles/colors.css"
+import { colors } from "styles/colors.css"
 
 interface SliderFieldCommonProps {
 	/** Identifies the field when a form is submitted. */
