+++ sanity/schemas/blog/blog-1/README.md
@@ -2,75 +2,28 @@
 
 This is a self-contained blog template. Each template has its own scoped Sanity schemas, app routes, and components. Multiple templates can coexist in the same project (e.g. a blog and a resources hub).
 
+The public URL for the blog is controlled entirely by the slug field in the Sanity hub singleton — no code changes are needed to set or change the route.
+
 ## Adopting this template for a project
 
-### 1. Choose a name
+### 1. Set the hub slug in Sanity Studio
 
-Decide on a short slug for this blog instance. Examples: `blog`, `resources`, `news`.
-This name is used throughout — in Sanity type names, route folders, and the hub slug.
+Go to Studio → Blog 1 → Hub Settings and set the slug to whatever the public URL should be (e.g. `blog`, `resources`, `news`). This drives the route, sitemap, and canonical URLs automatically.
 
-### 2. Rename the app route folder
-
-```
-app/blog-1/  →  app/blog/
-```
-
-All internal imports within the folder use relative paths so no changes are needed inside.
-
-### 3. Rename the Sanity schema folder
-
-```
-sanity/schemas/blog/blog-1/  →  sanity/schemas/blog/blog/
-```
-
-### 4. Rename Sanity type names
-
-In every file inside the schema folder, rename the type prefixes:
-
-| Old             | New (example)                                 |
-| --------------- | --------------------------------------------- |
-| `blog1Post`     | `post` (or `blogPost`, `resourcesPost`, etc.) |
-| `blog1Author`   | `author`                                      |
-| `blog1Category` | `category`                                    |
-| `blog1Hub`      | `blogHub`                                     |
-
-Files to update:
-
-- `postType.tsx` — `name: "blog1Post"` and any `to: [{ type: "blog1Post" }]` references
-- `authorType.tsx` — `name: "blog1Author"`
-- `categoryType.tsx` — `name: "blog1Category"`
-- `blockContentType.tsx` — `name: "blog1BlockContent"`
-- `publishWithReadTime.ts` — references `blog1Post`
-- `sanity/schemas/singletons/blog-1.tsx` — rename file and update `name: "blog1Hub"` and `to: [{ type: "blog1Post" }]`
-
-### 5. Update sanity.config.ts
-
-- Update imports to point to the renamed schema folder
-- Update the `blog1Types` array to use the new type names
-- Update the structure `blog1Item` to use the new type names and titles
-- Update `singletonPlugin([..., "blog1Hub"])` to use the new hub type name
-- Update `document.actions` if it references `"blog1Post"`
-
-### 6. Update the sitemap
-
-In `app/sitemap.ts`, update the GROQ query:
-
-```ts
-*[_type == "blog1Post" ...]  →  *[_type == "post" ...]
-```
-
-### 7. Update brand colors and fonts
+### 2. Update brand colors
 
 In `app/styles/colors.css.ts`, replace the `blog1*` placeholder color tokens with project brand colors.
 
-In `app/styles/text.ts`, replace `sampleMetrics` with the correct capsize metrics for the project font. Get metrics from [seek-oss.github.io/capsize](https://seek-oss.github.io/capsize/).
+### 3. Update the grid
 
-### 8. Set the hub slug in Sanity Studio
+In `app/(blog-templates)/(blog-template-1)/[blogSlug]/layout.tsx`, update the `makeResponsiveGrid` values to match the project's design grid (column count, gutter, margin, design widths).
 
-Go to Studio → Blog → Hub Settings and generate/set the slug to match the route (e.g. `blog`).
+### 4. Delete unused templates
+
+Delete any unused template folders under `app/(blog-templates)/` and `sanity/schemas/blog/`.
 
 ---
 
 ## Adding a second blog template
 
-Copy the entire `blog-1` folder, give it a new prefix (e.g. `resources`), and follow the steps above with that prefix. Both templates can live in the same project with separate routes, Sanity types, and Studio structure sections.
+Copy the entire `(blog-template-1)` folder under `app/(blog-templates)/` and give it a new name (e.g. `(blog-template-2)`). Update the absolute imports inside it to reflect the new folder name. Copy and rename the Sanity schema folder and type names using the steps above. Set a different slug in Studio for the second hub — each template's URL is independent.
