+++ tsconfig.json
@@ -10,7 +10,12 @@
 			"*": ["./*"],
 			"page": ["./app/[[...slug]]/page"],
 			"images/*": ["./app/images/*"],
-			"library/*": ["./library/*"]
+			"library/*": ["./library/*"],
+			"styles/*": ["./app/styles/*"],
+			"components/*": ["./app/components/*"],
+			"sections/*": ["./app/sections/*"],
+			"utils/*": ["./app/utils/*"],
+			"types/*": ["./app/types/*"]
 		}
 	},
 	"include": ["**/*.ts", "**/*.tsx", ".next/types/**/*.ts", ".next/dev/types/**/*.ts", "**/*.mts"],
